refactor(crypto): extract constantTimeStringEqual into shared crypto helper
